requirements.txt
Flask==2.2.3
gunicorn==20.1.0
Flask-Cors==5.0.1
Flask-Session==0.4.0
psycopg2-binary==2.9.5  # Agregar esto para PostgreSQL
werkzeug==2.2.2
